* new-style:
  update style
  update style
  rename filename
  update style
  rename filename
  cleanup
  update
  style
  cleanup some style
  restruct
  add compass to gemfile

+++ Rakefile
@@ -5,14 +5,14 @@ require 'jekyll'
 REPONAME = "ranmocy/ranmocy"
 ROOT_DIR  = Pathname.new('.').expand_path
 POSTS_DIR = Pathname.new("_posts").expand_path
+SITE_DIR = Pathname.new("_site").expand_path
 
 YAMLFrontMatter =<<FRONT
 ---
 layout: default
 published: true
-date: 22 January 2013
+date: 11 May 1991
 category: %s
-tags: [%s]
 ---
 
 FRONT
@@ -21,7 +21,8 @@ FRONT
 namespace :site do
   desc "Cleanup"
   task :cleanup do
-    FileUtils.rm_r(POSTS_DIR)
+    FileUtils.rm_r(SITE_DIR) if SITE_DIR.exist?
+    FileUtils.rm_r(POSTS_DIR) if POSTS_DIR.exist?
     FileUtils.mkdir(POSTS_DIR)
   end
 
@@ -29,13 +30,13 @@ namespace :site do
   task :prepare => [:cleanup] do
     Dir["[^_]*/*.*"].each do |file|
       ord_file = ROOT_DIR.join(file)
-      filename = ord_file.basename
+      filename = ord_file.basename.to_s.gsub(/\s+/, '-')
       category = ord_file.dirname.basename
-      new_file = POSTS_DIR.join("2013-04-13-#{filename}")
+      new_file = POSTS_DIR.join("1991-05-11-#{filename}")
 
       # FileUtils.cp(ord_file, new_file)
       File.open(new_file, "w") { |file|
-        file.puts (YAMLFrontMatter % [category, category])
+        file.puts (YAMLFrontMatter % [category])
         file.puts File.open(ord_file, "r").read
       }
     end
